app/equity_updater/requirements.txt
Django==3.2.6
psycopg2-binary==2.9.1
SQLAlchemy==1.4.35
investpy==1.0.8
pandas==1.5.2
apscheduler
finnhub-python==2.4.15
